+++ src/TextureCompositor.Thumbnails.cs
@@ -10,6 +10,8 @@ namespace DCFApixels.WhimTex
 
         internal Texture2D GetLayerThumbnail(Layer layer, int size, bool deferUpdates = false)
         {
+            if (layer?.Behaviour is FileLayerBehaviour file && file.sourceTexture != null)
+                return ResolveOriginalFileTexture(file.sourceTexture);
             if (layer?.Behaviour is TargetedLayerBehaviour || layer?.Behaviour is ShaderProcessorLayerBehaviour)
             {
                 layerThumbnails ??= new LayerThumbnailCache(this);